Signup: Verticals Survey: Shuffle categories when retrieved rather than during render
